Brought test/attest_test.rb up to date
Many sections have been commented out because the way Attest works
is not exactly the same as the way Dfect works.  These will be removed
in future, but it's important to explain these differences in the test
code before removing them.  See the comments for details.

All tests currently pass, but the Attest code that causes them to pass
is not currently committed.  See next commit...