Documentation; new feature Attest.caught_value
Attest.caught_value records the value (if any) most recently caught
by a C() or C?() method.  Use it for testing the correctness of that
value straigt after the C assertion.

Documentation: doc/attest.markdown (initial commit).  Comprehensive
manual to be hosted at gsinclair.github.com/attest.html.  Not finished
yet, but going pretty well.  This will be the master copy, but it will
be in the gsinclair.github.com repository as well.