Fail fast when error occurs; code clarity and comments
By rescuing ErrorOccurred in the correct place (in 'execute'), tests
no longer keep running after an error occurs.  This aids the testing
process by cutting down the number of errors/failures reported to you
at once.  (Arguably, this feature should be optional.)  Next: fail fast
when a failure occurs (not implemented in this commit).

In addition:
 * The execution process is heavily covered by debug messages.  They
   are colourful and indented.
 * The 'execute' method is heavily commented to get a handle of what's
   going on.
 * @suite changed to @current_suite (much clearer).
 * Local variable 'test' changed to 'assertion' in the 'action' method
   (much clearer).
 * New alias 'local_attest' to run the 'attest' command from the
   project directory, instead of from the installed gem.