Removed failing "insulation" tests
It was expected that certain things done inside an insulated test
(D!) would be invisible outside of that environment: class and
constant definitions. This was expected to be false for 1.8 but true
for 1.9. As of 1.9.3, it no longer appears to be true.

This is not an important feature for me, so I'm happy to remove the
failing tests and any mention of that "feature" in the documentation.